git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5520 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/SRD/fix_wall_srd.cpp
@@ -230,7 +230,7 @@ void FixWallSRD::wall_params(int flag)
 
   if (varflag) modify->clearstep_compute();
 
-  int ntimestep = update->ntimestep;
+  bigint ntimestep = update->ntimestep;
 
   for (int m = 0; m < nwall; m++) {
     if (wallstyle[m] == VARIABLE)
