+++ src/output.h
@@ -15,33 +15,34 @@
 #define LMP_OUTPUT_H
 
 #include "pointers.h"
+#include "lmptype.h"
 
 namespace LAMMPS_NS {
 
 class Output : protected Pointers {
  public:
-  int next;                    // next timestep for any kind of output
+  bigint next;                 // next timestep for any kind of output
 
-  int next_thermo;             // next timestep for thermo output
+  bigint next_thermo;          // next timestep for thermo output
   int thermo_every;            // thermo output every this many steps
-  int last_thermo;             // last timestep thermo was output
+  bigint last_thermo;          // last timestep thermo was output
   char *var_thermo;            // variable name for thermo frequency
   int ivar_thermo;             // variable index for thermo frequency
   class Thermo *thermo;        // Thermodynamic computations
 
   int ndump;                   // # of Dumps defined
   int max_dump;                // max size of Dump list
-  int next_dump_any;           // next timestep for any Dump
+  bigint next_dump_any;        // next timestep for any Dump
   int *every_dump;             // output of each Dump every this many steps
-  int *next_dump;              // next timestep to do each Dump
-  int *last_dump;              // last timestep each a snapshot was output
+  bigint *next_dump;           // next timestep to do each Dump
+  bigint *last_dump;           // last timestep each snapshot was output
   char **var_dump;             // variable name for dump frequency
   int *ivar_dump;              // variable index for dump frequency
   class Dump **dump;           // list of defined Dumps
 
-  int next_restart;            // next timestep to write a restart file
+  bigint next_restart;         // next timestep to write a restart file
   int restart_every;           // write a restart file every this many steps
-  int last_restart;            // last timestep a restart file was output
+  bigint last_restart;         // last timestep a restart file was output
   int restart_toggle;          // 0 if use restart1 as prefix
                                // 1 if use restart1 as file, 2 for restart2
   char *restart1,*restart2;    // names for restart files
@@ -51,9 +52,9 @@ class Output : protected Pointers {
   ~Output();
   void init();
   void setup(int);                   // initial output before run/min
-  void write(int);                   // output for current timestep
-  void write_dump(int);              // force output of dump snapshots
-  void write_restart(int);           // force output of a restart file
+  void write(bigint);                // output for current timestep
+  void write_dump(bigint);           // force output of dump snapshots
+  void write_restart(bigint);        // force output of a restart file
 
   void add_dump(int, char **);       // add a Dump to Dump list
   void modify_dump(int, char **);    // modify a Dump
